% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{abort_join_matches_remaining}
\alias{abort_join_matches_remaining}
\title{\code{unmatched = "error"}, y-side: "Each row of y must be matched by x"}
\usage{
abort_join_matches_remaining(right_unmatched)
}
\description{
Mirrors \code{dplyr:::rethrow_error_join_matches_remaining()} exactly.
}
\keyword{internal}
