% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{add_null_columns}
\alias{add_null_columns}
\title{Add null columns to a table}
\usage{
add_null_columns(tbl, col_names, col_types)
}
\arguments{
\item{tbl}{A tbl_gpu object}

\item{col_names}{Names of columns to add}

\item{col_types}{Types of columns to add}
}
\value{
A tbl_gpu with additional null columns
}
\description{
Add null columns to a table
}
\keyword{internal}
