% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{as_lazy}
\alias{as_lazy}
\title{Switch to lazy execution mode}
\usage{
as_lazy(.data)
}
\arguments{
\item{.data}{A \code{tbl_gpu} object.}
}
\value{
A \code{tbl_gpu} in lazy execution mode.
}
\description{
Returns a \code{tbl_gpu} that will defer operations until \code{collect()} or
\code{compute()} is called.
}
\examples{
if (has_gpu()) {
  # Create in eager mode, switch to lazy
  gpu_data <- tbl_gpu(mtcars) |>
    as_lazy() |>
    filter(mpg > 20) |>
    mutate(kpl = mpg * 0.425)  # not yet executed

  # Execute with collect
  result <- gpu_data |> collect()
}
}
