% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-gpu.R
\name{as_tbl_gpu}
\alias{as_tbl_gpu}
\title{Convert to tbl_gpu}
\usage{
as_tbl_gpu(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Additional arguments}
}
\value{
A tbl_gpu object
}
\description{
Convert to tbl_gpu
}
