% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_arrange}
\alias{ast_arrange}
\title{Create an arrange AST node}
\usage{
ast_arrange(input, sort_specs, groups = character())
}
\arguments{
\item{input}{Input AST node}

\item{sort_specs}{List of sort specifications (col_name, descending)}

\item{groups}{Character vector of group columns (for .by_group)}
}
\value{
An ast_arrange node
}
\description{
Create an arrange AST node
}
\keyword{internal}
