% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_barrier}
\alias{ast_barrier}
\title{Create a barrier AST node (optimization fence)}
\usage{
ast_barrier(input)
}
\arguments{
\item{input}{Input AST node}
}
\value{
An ast_barrier node
}
\description{
Create a barrier AST node (optimization fence)
}
\keyword{internal}
