% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_count}
\alias{ast_count}
\title{Count nodes in an AST tree}
\usage{
ast_count(node)
}
\arguments{
\item{node}{Root AST node}
}
\value{
Integer count
}
\description{
Count nodes in an AST tree
}
\keyword{internal}
