% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_depth}
\alias{ast_depth}
\title{Get the depth of an AST tree}
\usage{
ast_depth(node)
}
\arguments{
\item{node}{Root AST node}
}
\value{
Integer depth
}
\description{
Get the depth of an AST tree
}
\keyword{internal}
