% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_filter}
\alias{ast_filter}
\title{Create a filter AST node}
\usage{
ast_filter(input, predicates)
}
\arguments{
\item{input}{Input AST node}

\item{predicates}{List of predicate structures}
}
\value{
An ast_filter node
}
\description{
Create a filter AST node
}
\keyword{internal}
