% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_group_by}
\alias{ast_group_by}
\title{Create a group_by AST node (metadata only)}
\usage{
ast_group_by(input, groups)
}
\arguments{
\item{input}{Input AST node}

\item{groups}{Character vector of grouping columns}
}
\value{
An ast_group_by node
}
\description{
Create a group_by AST node (metadata only)
}
\keyword{internal}
