% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_head}
\alias{ast_head}
\title{Create a head/limit AST node}
\usage{
ast_head(input, n)
}
\arguments{
\item{input}{Input AST node}

\item{n}{Number of rows to keep}
}
\value{
An ast_head node
}
\description{
Create a head/limit AST node
}
\keyword{internal}
