% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_mutate}
\alias{ast_mutate}
\title{Create a mutate AST node}
\usage{
ast_mutate(input, expressions)
}
\arguments{
\item{input}{Input AST node}

\item{expressions}{List of expression structures}
}
\value{
An ast_mutate node
}
\description{
Create a mutate AST node
}
\keyword{internal}
