% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_node}
\alias{ast_node}
\title{Create an AST node}
\usage{
ast_node(type, ..., input = NULL)
}
\arguments{
\item{type}{Character string identifying the node type}

\item{...}{Additional named arguments stored on the node}

\item{input}{The input AST node (NULL for source nodes)}
}
\value{
An ast_node object
}
\description{
Create an AST node
}
\keyword{internal}
