% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_select}
\alias{ast_select}
\title{Create a select AST node}
\usage{
ast_select(input, columns)
}
\arguments{
\item{input}{Input AST node}

\item{columns}{Character vector of column names to keep}
}
\value{
An ast_select node
}
\description{
Create a select AST node
}
\keyword{internal}
