% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_source}
\alias{ast_source}
\title{Create a source AST node (leaf node)}
\usage{
ast_source(schema, source_ptr = NULL)
}
\arguments{
\item{schema}{List with names and types vectors}
}
\value{
An ast_source node
}
\description{
Create a source AST node (leaf node)
}
\keyword{internal}
