% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_ungroup}
\alias{ast_ungroup}
\title{Create an ungroup AST node (metadata only)}
\usage{
ast_ungroup(input)
}
\arguments{
\item{input}{Input AST node}
}
\value{
An ast_ungroup node
}
\description{
Create an ungroup AST node (metadata only)
}
\keyword{internal}
