% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{attach_ast_bottom}
\alias{attach_ast_bottom}
\title{Attach a node at the bottom of an AST}
\usage{
attach_ast_bottom(ast, new_input)
}
\arguments{
\item{ast}{AST node}

\item{new_input}{Node to attach at bottom}
}
\value{
AST with bottom input replaced
}
\description{
Returns a new AST with the bottom input replaced by \code{new_input}.
}
\keyword{internal}
