% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_cols_gpu}
\alias{bind_cols_gpu}
\title{GPU bind_cols implementation}
\usage{
bind_cols_gpu(dots, .name_repair = "unique")
}
\arguments{
\item{dots}{List of tables to bind}

\item{.name_repair}{Name repair method}
}
\value{
A tbl_gpu
}
\description{
GPU bind_cols implementation
}
\keyword{internal}
