% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_rows_gpu}
\alias{bind_rows_gpu}
\title{GPU bind_rows implementation}
\usage{
bind_rows_gpu(dots, .id = NULL)
}
\arguments{
\item{dots}{List of tables to bind}

\item{.id}{Optional column name to identify source}
}
\value{
A tbl_gpu
}
\description{
GPU bind_rows implementation
}
\keyword{internal}
