% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{check_relationship_many_to_one}
\alias{check_relationship_many_to_one}
\title{\code{relationship = "many-to-one"} violation: forbids an \code{x} row matching
multiple \code{y} rows (i.e. \code{left_multi_first}) -- ALWAYS reported as
"x matches multiple in y" (mirror image of one-to-many above).}
\usage{
check_relationship_many_to_one(left_multi)
}
\description{
\code{relationship = "many-to-one"} violation: forbids an \code{x} row matching
multiple \code{y} rows (i.e. \code{left_multi_first}) -- ALWAYS reported as
"x matches multiple in y" (mirror image of one-to-many above).
}
\keyword{internal}
