% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.tbl_gpu}
\alias{collect.tbl_gpu}
\title{Collect GPU table back to R}
\usage{
\method{collect}{tbl_gpu}(x, ...)
}
\arguments{
\item{x}{A tbl_gpu object}

\item{...}{Additional arguments (ignored)}
}
\value{
A tibble
}
\description{
Transfers data from GPU memory back to an R tibble.
}
\examples{
if (interactive()) {
  gpu_df <- tbl_gpu(mtcars)
  result <- gpu_df |> filter(mpg > 20) |> collect()
}
}
