% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{disable_cuda_stubs}
\alias{disable_cuda_stubs}
\title{Disable CUDA stub libraries in a directory}
\usage{
disable_cuda_stubs(lib_dir)
}
\arguments{
\item{lib_dir}{Path to search for stubs.}
}
\value{
Character vector of disabled stub paths (invisible).
}
\description{
Disable CUDA stub libraries in a directory
}
\keyword{internal}
