% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{extract_segment_above_barrier}
\alias{extract_segment_above_barrier}
\title{Extract segment above a barrier (returns copy)}
\usage{
extract_segment_above_barrier(ast, barrier)
}
\arguments{
\item{ast}{Root node}

\item{barrier}{Target barrier node}
}
\value{
Copy of segment with NULL input at barrier point
}
\description{
Extract segment above a barrier (returns copy)
}
\keyword{internal}
