% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{find_ast_bottom}
\alias{find_ast_bottom}
\title{Find the bottom node of an AST (deepest non-NULL input)}
\usage{
find_ast_bottom(ast)
}
\arguments{
\item{ast}{AST node}
}
\value{
Bottom node
}
\description{
Find the bottom node of an AST (deepest non-NULL input)
}
\keyword{internal}
