% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{find_calls}
\alias{find_calls}
\title{Recursively find all function calls in an R expression}
\usage{
find_calls(expr)
}
\arguments{
\item{expr}{An R expression (from parse())}
}
\value{
Character vector of function names
}
\description{
Recursively find all function calls in an R expression
}
\keyword{internal}
