% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{find_installed_cuplyr_so}
\alias{find_installed_cuplyr_so}
\title{Find installed cuplyr shared object path}
\usage{
find_installed_cuplyr_so()
}
\value{
Path to installed cuplyr shared object, or NULL if not found.
}
\description{
Find installed cuplyr shared object path
}
\keyword{internal}
