% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{find_next_barrier}
\alias{find_next_barrier}
\title{Find the next barrier node walking down from root}
\usage{
find_next_barrier(ast)
}
\arguments{
\item{ast}{AST node}
}
\value{
Barrier node or NULL
}
\description{
Find the next barrier node walking down from root
}
\keyword{internal}
