% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{fuse_filters}
\alias{fuse_filters}
\title{Mark filter nodes for fused AND-mask lowering}
\usage{
fuse_filters(ast)
}
\arguments{
\item{ast}{Root AST node}
}
\value{
AST with fused flags set
}
\description{
Mark filter nodes for fused AND-mask lowering
}
\keyword{internal}
