% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{fuse_mutates}
\alias{fuse_mutates}
\title{Fuse consecutive mutate nodes when safe}
\usage{
fuse_mutates(ast)
}
\arguments{
\item{ast}{Root AST node}
}
\value{
AST with fused mutates
}
\description{
Fuse consecutive mutate nodes when safe
}
\keyword{internal}
