% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_vars.tbl_gpu}
\alias{group_vars.tbl_gpu}
\title{Get grouping variables from a GPU table}
\usage{
\method{group_vars}{tbl_gpu}(x)
}
\arguments{
\item{x}{A \code{tbl_gpu} object.}
}
\value{
A character vector of grouping column names.
}
\description{
Returns the names of columns used for grouping.
}
\examples{
if (has_gpu()) {
  gpu_mtcars <- tbl_gpu(mtcars)

  grouped <- gpu_mtcars |>
    group_by(cyl, gear)

  group_vars(grouped)  # c("cyl", "gear")
}
}
