% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{groups.tbl_gpu}
\alias{groups.tbl_gpu}
\title{Get grouping information from a GPU table}
\usage{
\method{groups}{tbl_gpu}(x)
}
\arguments{
\item{x}{A \code{tbl_gpu} object.}
}
\value{
A list of symbols for the grouping columns.
}
\description{
Returns a list of symbols representing the grouping columns.
}
\examples{
if (has_gpu()) {
  gpu_mtcars <- tbl_gpu(mtcars)

  grouped <- gpu_mtcars |>
    group_by(cyl)

  groups(grouped)  # list(as.symbol("cyl"))
}
}
