% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{has_pending_ops}
\alias{has_pending_ops}
\title{Check if a tbl_gpu has pending lazy operations}
\usage{
has_pending_ops(.data)
}
\arguments{
\item{.data}{A \code{tbl_gpu} object.}
}
\value{
Logical. \code{TRUE} if there are pending operations.
}
\description{
Check if a tbl_gpu has pending lazy operations
}
