% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{infer_schema}
\alias{infer_schema}
\title{Infer output schema from an AST node}
\usage{
infer_schema(node)
}
\arguments{
\item{node}{An AST node}
}
\value{
List with names and types vectors
}
\description{
Infer output schema from an AST node
}
