% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_is_const}
\alias{ir_is_const}
\title{Is an IR subtree constant (no column references)?}
\usage{
ir_is_const(ir)
}
\arguments{
\item{ir}{An IR node}
}
\value{
\code{TRUE} iff \code{\link[=ir_cols]{ir_cols()}} is empty
}
\description{
Is an IR subtree constant (no column references)?
}
\keyword{internal}
