% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir_parse_recode_values}
\alias{ir_parse_recode_values}
\title{Parse \code{recode_values(x, old_values ~ new_value, ..., default=, unmatched=, ptype=)}}
\usage{
ir_parse_recode_values(expr, env, schema)
}
\description{
See \code{\link[=ir_parse_value_match_clauses]{ir_parse_value_match_clauses()}}. \verb{from=}/\verb{to=}/\verb{unmatched=}/\verb{ptype=}
are not implemented this wave.
}
\keyword{internal}
