% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{is_lazy}
\alias{is_lazy}
\title{Check if a tbl_gpu uses lazy execution}
\usage{
is_lazy(.data)
}
\arguments{
\item{.data}{A \code{tbl_gpu} object.}
}
\value{
Logical. \code{TRUE} if the table is in lazy execution mode.
}
\description{
Check if a tbl_gpu uses lazy execution
}
\examples{
if (has_gpu()) {
  eager_tbl <- tbl_gpu(mtcars)
  is_lazy(eager_tbl)  # FALSE

  lazy_tbl <- tbl_gpu(mtcars, lazy = TRUE)
  is_lazy(lazy_tbl)   # TRUE
}
}
