% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{is_opaque_expression}
\alias{is_opaque_expression}
\title{Check if an expression is opaque (contains unknown functions)}
\usage{
is_opaque_expression(expr_text)
}
\arguments{
\item{expr_text}{Expression as text string}
}
\value{
TRUE if expression contains unknown/unsafe functions
}
\description{
Check if an expression is opaque (contains unknown functions)
}
\keyword{internal}
