% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower_and_execute}
\alias{lower_and_execute}
\title{Lower and execute an AST against a GPU table}
\usage{
lower_and_execute(ast, source_ptr)
}
\arguments{
\item{ast}{Optimized AST root node}

\item{source_ptr}{External pointer to source GPU table}
}
\value{
External pointer to result GPU table
}
\description{
Lower and execute an AST against a GPU table
}
\keyword{internal}
