% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-gpu.R
\name{names<-.tbl_gpu}
\alias{names<-.tbl_gpu}
\title{Set column names of a GPU table}
\usage{
\method{names}{tbl_gpu}(x) <- value
}
\arguments{
\item{x}{A \code{tbl_gpu} object.}

\item{value}{A character vector of new column names. Must have the same
length as the number of columns, contain no \code{NA} values, and no empty
strings.}
}
\value{
The modified \code{tbl_gpu} object.
}
\description{
Set column names of a GPU table
}
