% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimize_with_barriers}
\alias{optimize_with_barriers}
\title{Optimize AST respecting barrier nodes}
\usage{
optimize_with_barriers(ast)
}
\arguments{
\item{ast}{Root AST node}
}
\value{
Optimized AST
}
\description{
Optimize AST respecting barrier nodes
}
\keyword{internal}
