% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{push_down_projections}
\alias{push_down_projections}
\title{Push down column projections to reduce data width early}
\usage{
push_down_projections(ast, required_cols = NULL, group_cols = character())
}
\arguments{
\item{ast}{Root AST node}

\item{required_cols}{Columns required by parent (NULL = all output cols)}

\item{group_cols}{Group columns that must be preserved}
}
\value{
Optimized AST with select nodes inserted
}
\description{
Push down column projections to reduce data width early
}
\keyword{internal}
