% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{repair_names}
\alias{repair_names}
\title{Repair duplicate column names}
\usage{
repair_names(names, method = "unique")
}
\arguments{
\item{names}{Character vector of column names}

\item{method}{Repair method}
}
\value{
Character vector with repaired names
}
\description{
Repair duplicate column names
}
\keyword{internal}
