% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{set_ast_source_ptr}
\alias{set_ast_source_ptr}
\title{Attach a source pointer to the source node in an AST}
\usage{
set_ast_source_ptr(node, ptr)
}
\arguments{
\item{node}{Root AST node}

\item{ptr}{External pointer to GPU table}
}
\value{
AST node with source pointer attached
}
\description{
Attach a source pointer to the source node in an AST
}
\keyword{internal}
