% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{set_op_is_factor}
\alias{set_op_is_factor}
\title{Is \code{name} a factor column in this schema?}
\usage{
set_op_is_factor(schema, name)
}
\description{
Is \code{name} a factor column in this schema?
}
\keyword{internal}
