% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{show_query}
\alias{show_query}
\title{Show pending operations for a lazy tbl_gpu}
\usage{
show_query(.data)
}
\arguments{
\item{.data}{A \code{tbl_gpu} object.}
}
\value{
Invisibly returns the AST, prints it for inspection.
}
\description{
Show pending operations for a lazy tbl_gpu
}
