% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{toposort_expressions}
\alias{toposort_expressions}
\title{Topologically sort expressions to respect dependencies}
\usage{
toposort_expressions(exprs)
}
\arguments{
\item{exprs}{List of expression structures}
}
\value{
Sorted list or NULL if cycle detected
}
\description{
Topologically sort expressions to respect dependencies
}
\keyword{internal}
