% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{try_fuse_mutate_pair}
\alias{try_fuse_mutate_pair}
\title{Try to fuse two consecutive mutate nodes}
\usage{
try_fuse_mutate_pair(lower, upper)
}
\arguments{
\item{lower}{Lower (earlier) mutate node}

\item{upper}{Upper (later) mutate node}
}
\value{
Fused mutate node or NULL if fusion not possible
}
\description{
Try to fuse two consecutive mutate nodes
}
\keyword{internal}
