% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{update_schema_after_mutate}
\alias{update_schema_after_mutate}
\title{Update schema after a mutate expression}
\usage{
update_schema_after_mutate(schema, expr)
}
\description{
Update schema after a mutate expression
}
\keyword{internal}
