% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{align_to_schema}
\alias{align_to_schema}
\title{Align a table to a target schema}
\usage{
align_to_schema(tbl, target_schema)
}
\arguments{
\item{tbl}{A tbl_gpu object}

\item{target_schema}{List with names and types}
}
\value{
A tbl_gpu aligned to the target schema
}
\description{
Align a table to a target schema
}
\keyword{internal}
