% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_join}
\alias{ast_join}
\title{Create a join AST node}
\usage{
ast_join(
  type,
  left,
  right,
  by,
  keep = FALSE,
  suffix = c(".x", ".y"),
  na_matches = "na"
)
}
\arguments{
\item{type}{Join type: "inner", "left", "right", "full"}

\item{left}{Left input AST node}

\item{right}{Right input AST node}

\item{by}{Join specification list(left = \if{html}{\out{<chr>}}, right = \if{html}{\out{<chr>}})}

\item{keep}{Logical, keep both key columns when names match}

\item{suffix}{Character vector of length 2}

\item{na_matches}{Character, "na" or "never"}
}
\value{
An ast_join node
}
\description{
Create a join AST node
}
\keyword{internal}
