% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast_summarise}
\alias{ast_summarise}
\title{Create a summarise AST node}
\usage{
ast_summarise(input, aggregations, groups)
}
\arguments{
\item{input}{Input AST node}

\item{aggregations}{List of aggregation structures}

\item{groups}{Character vector of group columns (stored explicitly)}
}
\value{
An ast_summarise node
}
\description{
Create a summarise AST node
}
\keyword{internal}
