% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{cast_column}
\alias{cast_column}
\title{Cast a column to a different type}
\usage{
cast_column(tbl, col_name, target_type)
}
\arguments{
\item{tbl}{A tbl_gpu object}

\item{col_name}{Name of column to cast}

\item{target_type}{Target type string}
}
\value{
A tbl_gpu with the column cast to the target type
}
\description{
Cast a column to a different type
}
\keyword{internal}
