% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{compute_unified_schema}
\alias{compute_unified_schema}
\title{Compute unified schema from multiple tables}
\usage{
compute_unified_schema(tables)
}
\arguments{
\item{tables}{List of tbl_gpu objects}
}
\value{
List with names and types for unified schema
}
\description{
Compute unified schema from multiple tables
}
\keyword{internal}
