% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{infer_mutate_output_type}
\alias{infer_mutate_output_type}
\title{Infer output type for a mutate expression}
\usage{
infer_mutate_output_type(op, input_types, scalar)
}
\arguments{
\item{op}{The operation}

\item{input_types}{Types of input columns}

\item{scalar}{Scalar value if any}
}
\value{
GPU type string
}
\description{
Infer output type for a mutate expression
}
\keyword{internal}
