% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{is_barrier}
\alias{is_barrier}
\title{Check if an AST node is an optimization barrier}
\usage{
is_barrier(node)
}
\arguments{
\item{node}{An AST node}
}
\value{
TRUE if node is a barrier
}
\description{
Check if an AST node is an optimization barrier
}
\keyword{internal}
