% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{make_aggregation}
\alias{make_aggregation}
\title{Create an aggregation structure for summarise}
\usage{
make_aggregation(output_col, input_col, fn, input_type = NULL)
}
\arguments{
\item{output_col}{Output column name}

\item{input_col}{Input column name}

\item{fn}{Aggregation function name (sum, mean, min, max, n)}

\item{input_type}{Input column type}
}
\value{
An aggregation list structure
}
\description{
Create an aggregation structure for summarise
}
\keyword{internal}
