% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{make_mutate_expr}
\alias{make_mutate_expr}
\title{Create a mutate expression structure}
\usage{
make_mutate_expr(output_col, input_cols, op, scalar = NULL, input_types = NULL)
}
\arguments{
\item{output_col}{Output column name}

\item{input_cols}{Character vector of input column names}

\item{op}{Operation: "+", "-", "*", "/", "^", "copy", or function name}

\item{scalar}{Numeric scalar or NULL}

\item{input_types}{Character vector of input column types}
}
\value{
An expression list structure
}
\description{
Create a mutate expression structure
}
\keyword{internal}
