% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{make_predicate}
\alias{make_predicate}
\title{Create a filter predicate structure}
\usage{
make_predicate(col_name, op, value, is_col_compare = FALSE)
}
\arguments{
\item{col_name}{Column name for LHS}

\item{op}{Comparison operator}

\item{value}{Scalar value or column name for RHS}

\item{is_col_compare}{TRUE if RHS is a column name}
}
\value{
A predicate list structure
}
\description{
Create a filter predicate structure
}
\keyword{internal}
