% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate.tbl_gpu}
\alias{mutate.tbl_gpu}
\title{Create or modify columns in a GPU table}
\usage{
\method{mutate}{tbl_gpu}(.data, ...)
}
\arguments{
\item{.data}{A tbl_gpu object}

\item{...}{Name-value pairs of expressions (e.g., z = x + y)}
}
\value{
A tbl_gpu with new/modified columns
}
\description{
Create or modify columns in a GPU table
}
\examples{
if (interactive()) {
  gpu_df <- tbl_gpu(mtcars)
  gpu_df |> mutate(kpl = mpg * 0.425)
}
}
