% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimize_segment}
\alias{optimize_segment}
\title{Optimize a single segment (no barriers)}
\usage{
optimize_segment(ast)
}
\arguments{
\item{ast}{Root of segment}
}
\value{
Optimized segment
}
\description{
Optimize a single segment (no barriers)
}
\keyword{internal}
