% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{patch_makevars_for_cloud}
\alias{patch_makevars_for_cloud}
\title{Patch Makevars for Colab/cloud: add conda lib and driver to RUNPATH}
\usage{
patch_makevars_for_cloud(src_dir, driver_lib)
}
\arguments{
\item{src_dir}{Source directory containing src/Makevars.}

\item{driver_lib}{Real driver library path.}
}
\description{
Patch Makevars for Colab/cloud: add conda lib and driver to RUNPATH
}
\keyword{internal}
